+++ packages/core/styles/src/types/index.ts
@@ -69,3 +69,11 @@ export type TransitionTiming =
   | "step-start"
   | "step-end";
 export type WordBreak = "normal" | "break-all" | "keep-all" | "break-word";
+export type TextOverflow = "clip" | "ellipsis";
+export type WhiteSpace =
+  | "normal"
+  | "nowrap"
+  | "pre"
+  | "pre-wrap"
+  | "pre-line"
+  | "break-spaces";