+++ packages/react/src/atomic/Text/src/Text.tsx
@@ -14,6 +14,9 @@ const Text: React.FC<TextProps> & TextComponents = ({
   fontWeight = "regular",
   fontSize = "base",
   textDecoration = "none",
+  textOverflow,
+  whiteSpace,
+  overflow,
   children,
   lineClamp,
   ...rest
@@ -26,6 +29,9 @@ const Text: React.FC<TextProps> & TextComponents = ({
     fontWeight: fontWeight as any,
     fontSize: fontSize as any,
     textDecoration: textDecoration as any,
+    textOverflow,
+    whiteSpace,
+    overflow,
     WebkitLineClamp: lineClamp as any,
   });